Checkpoint - still working on redirect, adding TS support methods.
