+++ plugin/src/Context.cc
@@ -29,10 +29,10 @@ using namespace swoc::literals;
 
 Context::Context(Config & cfg) {
   // This is arranged so @a _arena destructor will clean up properly, nothing more need be done.
-  _arena.reset(swoc::MemArena::construct_self_contained(4000));
+  _arena.reset(swoc::MemArena::construct_self_contained(4000 + cfg._ctx_storage_required));
 
   // Provide array storage for each potential conditional directive for each hook.
-  for ( unsigned idx = static_cast<unsigned>(Hook::BEGIN) ; idx < static_cast<unsigned>(Hook::END) ; ++idx) {
+  for (unsigned idx = static_cast<unsigned>(Hook::BEGIN) ; idx < static_cast<unsigned>(Hook::END) ; ++idx) {
      MemSpan<Directive*> drtv_list = _arena->alloc(sizeof(Directive*) * cfg._directive_count[idx]).rebind<Directive*>();
      _directives[idx] = { 0, 0, drtv_list.data() };
   };
@@ -42,6 +42,9 @@ Context::Context(Config & cfg) {
   , [] (void*, void*) -> void {}, this);
   _rxp_capture = pcre2_match_data_create(cfg._capture_groups, _rxp_ctx);
   _rxp_working = pcre2_match_data_create(cfg._capture_groups, _rxp_ctx);
+
+  // Directive shared storage
+  _ctx_store = _arena->alloc(cfg._ctx_storage_required);
 }
 
 Errata Context::on_hook_do(Hook hook_idx, Directive *drtv) {
@@ -110,6 +113,13 @@ Context& Context::commit(FeatureData const &feature) {
   return *this;
 }
 
+swoc::MemSpan<void> Context::storage_for(Directive *drtv) {
+  auto zret { _ctx_store };
+  zret.remove_prefix(drtv->_rtti->_ctx_storage_offset);
+  zret.remove_suffix(zret.size() - drtv->_rtti->_ctx_storage_offset);
+  return zret;
+}
+
 ts::HttpHeader Context::creq_hdr() {
   if (!_creq.is_valid()) {
     _creq = _txn.creq_hdr();
@@ -122,6 +132,20 @@ ts::HttpHeader Context::preq_hdr() {
     _preq = _txn.preq_hdr();
   }
   return _preq;
+}
+
+ts::HttpHeader Context::ursp_hdr() {
+  if (!_ursp.is_valid()) {
+    _ursp = _txn.ursp_hdr();
+  }
+  return _ursp;
+}
+
+ts::HttpHeader Context::prsp_hdr() {
+  if (!_prsp.is_valid()) {
+    _prsp = _txn.prsp_hdr();
+  }
+  return _prsp;
 }
 
 unsigned Context::ArgPack::count() const {
