+++ plugin/src/Directive.cc
@@ -24,6 +24,9 @@ using swoc::Rv;
 using swoc::TextView;
 
 /* ------------------------------------------------------------------------------------ */
+const std::string Directive::DO_KEY { "do" };
+unsigned Directive::StaticInfo::_counter = 0;
+
 DirectiveList& DirectiveList::push_back(Directive::Handle &&d) {
   _directives.emplace_back(std::move(d));
   return *this;
